+++ scripts/migration/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @docknetwork/migration
 
+## 0.8.0
+
+### Minor Changes
+
+- Align with the latest `cheqd` node version (`3.1.1`)
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/cheqd-blockchain-api@0.18.0
+  - @docknetwork/credential-sdk@0.23.0
+  - @docknetwork/dock-blockchain-api@0.9.1
+  - @docknetwork/dock-blockchain-modules@0.12.1
+
 ## 0.7.0
 
 ### Minor Changes